feat: display productTs&Cs on checkout
