+++ support-frontend/assets/helpers/productCatalog.ts
@@ -1,7 +1,7 @@
 import type { ActiveProductKey } from '@guardian/support-service-lambdas/modules/product-catalog/src/productCatalog';
 import { activeTypeObject } from '@guardian/support-service-lambdas/modules/product-catalog/src/typeObject';
 import type { Participations } from './abTests/models';
-import type { RegularContributionType } from './contributions';
+import type { ContributionType } from './contributions';
 import { newspaperCountries } from './internationalisation/country';
 import type {
 	CountryGroupId,
@@ -206,8 +206,11 @@ const guardianAdLiteBenefits = [
 	{ copy: 'You can cancel at any time.' },
 ];
 
-const proceedingTsAndCs = {
+const proceedingTsAndCsDigitalPlusPrint = {
 	copy: 'By proceeding, you are agreeing to the Digital + print Terms.',
+};
+const proceedingTsAndCs = {
+	copy: 'By proceeding, you are agreeing to our Terms and Conditions.',
 };
 const personalDataTsAndCs = {
 	copy: 'To find out what personal data we collect and how we use it, please visit our Privacy Policy.',
@@ -223,28 +226,36 @@ const contributionTsAndCs = [
 ];
 const supporterPlusTsAndCs = [
 	{
-		copy: `If you pay at least £X per month, you will receive the All-access digital benefits on a subscription basis. If you increase your payments per month, these additional amounts will be separate monthly voluntary financial contributions to the Guardian. The All-access digital subscription and any contributions will auto-renew each month. You will be charged the subscription and contribution amounts using your chosen payment method at each renewal unless you cancel. You can cancel your subscription or change your contributions at any time before your next renewal date. If you cancel within 14 days of taking out a All-access digital subscription, you’ll receive a full refund (including of any contributions) and your subscription and any contribution will stop immediately. Cancellation of your subscription (which will also cancel any contribution) or cancellation of your contribution made after 14 days will take effect at the end of your current monthly payment period. To cancel, go to Manage My Account or see our Terms.`,
+		copy: `SUPPORTERPLUS: If you pay at least £X per month, you will receive the All-access digital benefits on a subscription basis. If you increase your payments per month, these additional amounts will be separate monthly voluntary financial contributions to the Guardian. The All-access digital subscription and any contributions will auto-renew each month. You will be charged the subscription and contribution amounts using your chosen payment method at each renewal unless you cancel. You can cancel your subscription or change your contributions at any time before your next renewal date. If you cancel within 14 days of taking out a All-access digital subscription, you’ll receive a full refund (including of any contributions) and your subscription and any contribution will stop immediately. Cancellation of your subscription (which will also cancel any contribution) or cancellation of your contribution made after 14 days will take effect at the end of your current monthly payment period. To cancel, go to Manage My Account or see our Terms.`,
 		promotionalCopy: `If you pay £XX per month for the first X months, then , then £X per month, you will receive the All-access digital benefits on a subscription basis. If you increase your payments per month, these additional amounts will be separate monthly voluntary financial contributions to the Guardian. The All-access digital subscription and any contributions will auto-renew each month. You will be charged the subscription and contribution amounts using your chosen payment method at each renewal unless you cancel. You can cancel your subscription or change your contributions at any time before your next renewal date. If you cancel within 14 days of taking out a All-access digital subscription, you’ll receive a full refund (including of any contributions) and your subscription and any contribution will stop immediately. Cancellation of your subscription (which will also cancel any contribution) or cancellation of your contribution made after 14 days will take effect at the end of your current monthly payment period. To cancel, go to Manage My Account or see our Terms.`,
 	},
-	...contributionTsAndCs,
-];
-
-const tierThreeTsAndCs = [
-	{
-		copy: `By signing up, you are taking out a Digital + print subscription. Your Digital + print subscription will auto-renew each month unless cancelled. Your first payment will be taken on the publication date of your first Guardian Weekly magazine (as shown in the checkout) but you will start to receive your digital benefits when you sign up. Unless you cancel, subsequent monthly payments will be taken on this date using your chosen payment method. You can cancel your Digital + print subscription at any time before your next renewal date. If you cancel your Digital + print subscription within 14 days of signing up, your subscription will stop immediately and we will not take the first payment from you. Cancellation of your subscription after 14 days will take effect at the end of your current monthly payment period. To cancel go to Manage My Account or see our Digital + print Terms.`,
-	},
-	{
-		copy: `By proceeding, you are agreeing to the Digital + print Terms.`,
-	},
+	proceedingTsAndCs,
 	personalDataTsAndCs,
 	stripeTsAndCs,
 ];
-
+const tierThreeTsAndCs = [
+	{
+		copy: `TIERTHREE: By signing up, you are taking out a Digital + print subscription. Your Digital + print subscription will auto-renew each month unless cancelled. Your first payment will be taken on the publication date of your first Guardian Weekly magazine (as shown in the checkout) but you will start to receive your digital benefits when you sign up. Unless you cancel, subsequent monthly payments will be taken on this date using your chosen payment method. You can cancel your Digital + print subscription at any time before your next renewal date. If you cancel your Digital + print subscription within 14 days of signing up, your subscription will stop immediately and we will not take the first payment from you. Cancellation of your subscription after 14 days will take effect at the end of your current monthly payment period. To cancel go to Manage My Account or see our Digital + print Terms.`,
+	},
+	proceedingTsAndCsDigitalPlusPrint,
+	personalDataTsAndCs,
+	stripeTsAndCs,
+];
 const adLiteTsAndCs = [
 	{
-		copy: `Your Guardian Ad-Lite subscription will auto-renew each month unless cancelled. Your first payment will be taken on day 15 after signing up but you will start to receive your Guardian Ad-Lite benefits when you sign up. Unless you cancel, subsequent monthly payments will be taken on this date using your chosen payment method. You can cancel your subscription at any time before your next renewal date. If you cancel your Guardian Ad-Lite subscription within 14 days of signing up, your subscription will stop immediately and we will not take the first payment from you. Cancellation of your subscription after 14 days will take effect at the end of your current monthly payment period. To cancel, go to Manage My Account or see our Guardian Ad-Lite Terms.`,
+		copy: `ADLITE: Your Guardian Ad-Lite subscription will auto-renew each month unless cancelled. Your first payment will be taken on day 15 after signing up but you will start to receive your Guardian Ad-Lite benefits when you sign up. Unless you cancel, subsequent monthly payments will be taken on this date using your chosen payment method. You can cancel your subscription at any time before your next renewal date. If you cancel your Guardian Ad-Lite subscription within 14 days of signing up, your subscription will stop immediately and we will not take the first payment from you. Cancellation of your subscription after 14 days will take effect at the end of your current monthly payment period. To cancel, go to Manage My Account or see our Guardian Ad-Lite Terms.`,
 	},
-	...contributionTsAndCs,
+	proceedingTsAndCs,
+	personalDataTsAndCs,
+	stripeTsAndCs,
+];
+const digitalEditionTsAndCs = [
+	{
+		copy: `DIGITALEDITION: Payment taken after the first 14 day free trial. At the end of the free trial period your subscription will auto-renew, and you will be charged, each month at the full price of £14.99 per month or £149 per year unless you cancel. You can cancel at any time before your next renewal date. Cancellation will take effect at the end of your current subscription month. To cancel, go to Manage My Account or see our Terms.`,
+	},
+	proceedingTsAndCs,
+	personalDataTsAndCs,
+	stripeTsAndCs,
 ];
 
 export const productCatalogDescription: Record<
@@ -316,11 +327,7 @@ export const productCatalogDescription: Record<
 				copy: 'Free 14 day trial. Enjoy a free trial of your subscription, before you pay',
 			},
 		],
-		tsAndCs: [
-			{
-				copy: 'Ts&Cs DigitalSubscription',
-			},
-		],
+		tsAndCs: digitalEditionTsAndCs,
 		ratePlans: {
 			Monthly: {
 				billingPeriod: 'Monthly',
@@ -603,7 +610,7 @@ export function productCatalogGuardianAdLite(): Record<
 export function productCatalogTsAndCs(
 	productKey: ActiveProductKey,
 	countryGroupId: CountryGroupId,
-	contributionType: RegularContributionType,
+	contributionType: ContributionType,
 	promotion?: Promotion,
 ): ProductTsAndCs[] {
 	const isoCurrency = detect(countryGroupId);
