+++ support-frontend/assets/pages/[countryGroupId]/components/checkoutComponent.tsx
@@ -90,12 +90,10 @@ import { logException } from 'helpers/utilities/logger';
 import type { GeoId } from 'pages/geoIdConfig';
 import { getGeoIdConfig } from 'pages/geoIdConfig';
 import { CheckoutDivider } from 'pages/supporter-plus-landing/components/checkoutDivider';
+import { CheckoutTsAndCs } from 'pages/supporter-plus-landing/components/checkoutTsAndCs';
 import { GuardianTsAndCs } from 'pages/supporter-plus-landing/components/guardianTsAndCs';
 import { PatronsMessage } from 'pages/supporter-plus-landing/components/patronsMessage';
-import {
-	PaymentTsAndCs,
-	SummaryTsAndCs,
-} from 'pages/supporter-plus-landing/components/paymentTsAndCs';
+import { SummaryTsAndCs } from 'pages/supporter-plus-landing/components/paymentTsAndCs';
 import {
 	formatMachineDate,
 	formatUserDate,
@@ -1414,7 +1412,20 @@ export function CheckoutComponent({
 								/>
 							</div>
 						)}
-						<PaymentTsAndCs
+						<CheckoutTsAndCs
+							productKey={productKey}
+							countryGroupId={countryGroupId}
+							contributionType={
+								productFields.billingPeriod === 'Monthly'
+									? 'MONTHLY'
+									: productFields.billingPeriod === 'Annual'
+									? 'ANNUAL'
+									: 'ONE_OFF'
+							}
+							promotion={promotion}
+						/>
+
+						{/* <PaymentTsAndCs />
 							countryGroupId={countryGroupId}
 							contributionType={
 								productFields.billingPeriod === 'Monthly'
@@ -1430,7 +1441,7 @@ export function CheckoutComponent({
 							}
 							productKey={productKey}
 							promotion={promotion}
-						/>
+						/> */}
 					</BoxContents>
 				</Box>
 			</form>
